Consolidate new tests into existing feature test files
Fold the definite-assignment/nullable/stackalloc/expression-tree tests into
LabeledBreakContinueBindingTests and the PDB tests into
LabeledBreakContinueEmitTests, removing the separate Flow/Pdb files (the
compiler team prefers not to add more test files).
